pyproject.toml
[project]
name = "multimodal-audit-engine"
version = "0.1.0"
description = "Video content compliance auditing system using multimodal AI analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.0",
    "langchain-core>=1.2.0",
    "langchain-community>=0.4.0",
    "langchain-mistralai>=1.1.0",
    "opensearch-py>=3.1.0",
    "yt-dlp>=2026.0.0",
    "pypdf>=5.4.0",
    "langchain-text-splitters>=1.1.0",
    "tiktoken>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "langsmith>=0.7.0",
    "opentelemetry-sdk>=1.39.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "openai-whisper>=20240626",
    "langgraph>=0.1.0",
    "faiss-cpu>=1.8.0",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
]
